Adding helper functions for clustering / dup. detection
